Revert "nazanin/amount_design"
