--- src/javascript/app_2/App/Components/Form/InputField/index.js
@@ -1,1 +1,0 @@
-export default from './input-field.jsx';
