+++ src/javascript/app_2/App/Components/Form/TimePicker/time-picker.jsx
@@ -4,8 +4,8 @@ import PropTypes         from 'prop-types';
 import React             from 'react';
 import { CSSTransition } from 'react-transition-group';
 import Dialog            from './dialog.jsx';
-import InputField        from '../InputField';
-import { IconClock }     from '../../../../Assets/Common/icon-clock.jsx';
+import InputField        from '../input-field.jsx';
+import { IconClock } from '../../../../Assets/Common/icon-clock.jsx';
 
 class TimePicker extends React.Component {
     state = { is_open: false };
