+++ src/javascript/app_2/Modules/Trading/Components/Form/TradeParams/allow-equals.jsx
@@ -2,7 +2,7 @@ import React          from 'react';
 import PropTypes      from 'prop-types';
 import { localize }   from '_common/localize';
 import Tooltip        from 'App/Components/Elements/tooltip.jsx';
-import InputField     from 'App/Components/Form/InputField';
+import InputField     from 'App/Components/Form/input-field.jsx';
 
 const AllowEquals = ({
     is_allow_equal,
