+++ src/javascript/app_2/Modules/Trading/Components/Form/TradeParams/barrier.jsx
@@ -6,8 +6,7 @@ import {
     IconBarrierUp,
     IconBarrierDown } from 'Assets/Trading/Barriers';
 import Fieldset       from 'App/Components/Form/fieldset.jsx';
-import InputField     from 'App/Components/Form/InputField';
-
+import InputField     from 'App/Components/Form/input-field.jsx';
 import { localize }   from '_common/localize';
 
 const Barrier = ({
