+++ src/templates/static/new_layout/style_guide.jsx
@@ -1,7 +1,7 @@
 import React      from 'react';
 import Tooltip    from '../../../javascript/app_2/App/Components/Elements/tooltip.jsx';
 import Button     from '../../../javascript/app_2/App/Components/Form/button.jsx';
-import InputField from '../../../javascript/app_2/App/Components/Form/InputField';
+import InputField from '../../../javascript/app_2/App/Components/Form/input-field.jsx';
 
 const StyleGuide = () => (
     <div className='container'>
